moved to training.cli package
